:bug: Update data types in communication structures and improve document formatting

+++ src/shared/include/Player.hpp
@@ -5,7 +5,7 @@
 typedef struct player {
     int id;
     char username[20];
-    double y_pos;
+    float y_pos;
     int coins_collected;
     bool is_dead;
     bool is_jetpack_on;
